# Работа с сетью и веб-сервисами/internal/models/task.go
package models

// Task describes a single to-do item returned by the API.
type Task struct {
	ID        int    `json:"id"`
	Title     string `json:"title"`
	Done      bool   `json:"done"`
	CreatedAt string `json:"created_at,omitempty"`
}